requirements.txt
fastapi>=0.111
uvicorn>=0.30
pydantic>=2.8
PyYAML>=6.0 ; python_version >= "3.8"
